pyproject.toml
[tool.poetry]
name = "task-management-api"
version = "0.1.0"
description = "FastAPI application for task management with background processing"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.22"
asyncpg = "^0.28.0"
alembic = "^1.12.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
